# [REVIEW]Hashing Vectorizer and general vectorizer improvements