--- a/python/cuml/common/sparsefuncs.py
+++ b/python/cuml/common/sparsefuncs.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 import math
-
+import cupy as cp
 from cuml.common import with_cupy_rmm
 from cuml.common.kernel_utils import cuda_kernel_factory
 
@@ -97,3 +97,59 @@ def csr_row_normalize_l2(X, inplace=True):
            (X.data, X.indices, X.indptr, X.shape[0]))
 
     return X
+
+
+def create_csr_matrix_from_count_df(count_df, empty_doc_ids, n_doc, n_features,
+                                    dtype=cp.float32):
+    """
+    Create a sparse matrix from the count of tokens by document
+
+    Parameters
+    ----------
+        count_df = cudf.DataFrame({'count':..., 'doc_id':.., 'token':.. })
+                    sorted by doc_id and token
+        empty_doc_ids = cupy array containing doc_ids with no tokens
+        n_doc: Total number of documents
+        n_features: Number of features
+        dtype: Output dtype
+    """
+    data = count_df["count"].values
+    indices = count_df["token"].values
+
+    doc_token_counts = count_df["doc_id"].value_counts().reset_index()
+    del count_df
+    doc_token_counts = doc_token_counts.rename(
+        {"doc_id": "token_counts", "index": "doc_id"}, axis=1
+    ).sort_values(by="doc_id")
+
+    token_counts = _insert_zeros(
+        doc_token_counts["token_counts"], empty_doc_ids
+    )
+    indptr = token_counts.cumsum()
+    indptr = cp.pad(indptr, (1, 0), "constant")
+
+    return cp.sparse.csr_matrix(
+        arg1=(data, indices, indptr), dtype=dtype,
+        shape=(n_doc, n_features)
+    )
+
+
+def _insert_zeros(ary, zero_indices):
+    """
+    Create a new array of len(ary + zero_indices) where zero_indices
+    indicates indexes of 0s in the new array. Ary is used to fill the rest.
+
+    Example:
+        _insert_zeros([1, 2, 3], [1, 3]) => [1, 0, 2, 0, 3]
+    """
+    if len(zero_indices) == 0:
+        return ary.values
+
+    new_ary = cp.zeros((len(ary) + len(zero_indices)), dtype=cp.int32)
+
+    # getting mask of non-zeros
+    data_mask = ~cp.in1d(cp.arange(0, len(new_ary), dtype=cp.int32),
+                         zero_indices)
+
+    new_ary[data_mask] = ary
+    return new_ary