… docs/source/api.rst
… python/cuml/feature_extraction/text.py
… python/cuml/test/test_text_feature_extraction.py