@VibhuJawa 2020-07-17
Just linking the test/benchmarking notebooks here:

- Doc Search using [Hash-Vectorizer](https://gist.github.com/VibhuJawa/d14dc19dec549cb00499920b6c0ece84) (Gets relevant same results as count-vectorizer as on the blog) 

- [PR performance Notebook](https://gist.github.com/VibhuJawa/34185c74605e918a958d999efefd807d)

- [Master Performance Notebook](https://gist.github.com/VibhuJawa/cdeedbf80bfa8dc68b1138e786d6ed6b)